+++ js/main.js
@@ -1,6 +1,10 @@
 import {createOffers} from './data.js';
 import {insertOffer} from './popup.js';
 import {getRandomPositiveInteger} from './util.js';
+import { activateForm, deactivateForm } from './map-form.js';
+
+deactivateForm();
+activateForm();
 
 const offers = createOffers();
 const randomIndex = getRandomPositiveInteger(0,9);
